New file...
Scrabble is a game where players get points by spelling words. Words
are scored by adding together the point values of each individual
letter (we'll leave out the double and triple letter and word scores
for now).

To the right is a dictionary containing all of the letters in the
alphabet with their corresponding Scrabble point values.

For example: the word "Helix" would score 15 points due to the sum of
the letters: 4 + 1 + 1 + 1 + 8.